requirements.txt
flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
open3d==0.17.0
pillow==10.0.0
transformers==4.33.2
timm==0.9.7